PR Feedback move Agent to <AppConfig/> component
