+++ packages/pwa/app/commerce-api/index.js
@@ -10,36 +10,9 @@ import * as sdk from 'commerce-sdk-isomorphic'
 import {getAppOrigin} from 'pwa-kit-react-sdk/utils/url'
 import OcapiShopperBaskets from './ocapi-shopper-baskets'
 import OcapiShopperOrders from './ocapi-shopper-orders'
-import {getTenantId, isServer, isError, isTokenValid} from './utils'
+import {getTenantId, isError, isTokenValid} from './utils'
 import Auth from './auth'
 import EinsteinAPI from './einstein'
-import http from 'http'
-import https from 'https'
-
-const KEEP_ALIVE_MS = 8000
-
-const HTTP_AGENT = isServer
-    ? new http.Agent({
-          keepAlive: true,
-          keepAliveMsecs: KEEP_ALIVE_MS
-      })
-    : undefined
-const HTTPS_AGENT = isServer
-    ? new https.Agent({
-          keepAlive: true,
-          keepAliveMsecs: KEEP_ALIVE_MS
-      })
-    : undefined
-
-/**
- * Node.js http or https Agent with keepAlive support to re-use Commerce API
- * connections server-side.
- * @param url
- * @returns {module:http.Agent|module:https.Agent}
- */
-function getAgent(url) {
-    return url.protocol === 'http:' ? HTTP_AGENT : HTTPS_AGENT
-}
 
 /**
  * The configuration details for the connecting to the API.
@@ -82,13 +55,7 @@ class CommerceAPI {
         // Client-side requests should be proxied via the configured path.
         const proxy = `${getAppOrigin()}${proxyPath}`
 
-        this._config = {
-            proxy,
-            fetchOptions: {
-                agent: getAgent
-            },
-            ...restConfig
-        }
+        this._config = {proxy, ...restConfig}
 
         this.auth = new Auth(this)
